fixed table report breaking on short host names (fixes #91)
There is a function meant to strip off the table part of the select
fields, and it works based on number of . characters in the column name.
However when we put hostnames in there it tends not to do the right thing.
It could use more refactoring, but this quick fix works for now.