fixed default time zone handling when date.timezone not set in php
Previous behavior was to force the timezone to be America/Los_Angeles
when none was set.  This caused issues because the data
in the table didn't agree with the offsets in the interface
and you would get incorrect query results

Now the default is to use the system timezone, which produces
more reasonable results.