added top n option to graph search ( fixes #115)
It would be useful to visualize just how much each of the top queries contribute to the aggregated values on the graph. We already have support to pivot based on hostnames, adding a pivot based on checksums would take a little more work.

The major change is that the current search page can fire ajax requests for both the table and graph result data simultaneously because they have no particular dependencies on each other. To get the top N checksums we need the data from the table result first.

This change will have to enable that -- when this checkbox is checked, we'll delay getting the graph results until the table results are returned, then use them to get the list of checksum values. Once we have those we can send the right data to the graph ajax request to make this happen.